// scripts/types/database/public/EspeceProtegee.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.espece_protegee */
export type EspeceProtegeeCdRef = string & { __brand: 'public.espece_protegee' };

/** Represents the table public.espece_protegee */
export default interface EspeceProtegee {
  cd_ref: EspeceProtegeeCdRef;

  classification: string;

  noms_scientifiques: string[];

  noms_vernaculaires: string[];

  cd_type_statuts: string[];

  espece_ministerielle: boolean;

  espece_cnpn: boolean;
}

/** Represents the initializer for the table public.espece_protegee */
export interface EspeceProtegeeInitializer {
  cd_ref: EspeceProtegeeCdRef;

  classification: string;

  /** Default value: '{}'::text[] */
  noms_scientifiques?: string[];

  /** Default value: '{}'::text[] */
  noms_vernaculaires?: string[];

  /** Default value: '{}'::text[] */
  cd_type_statuts?: string[];

  /** Default value: false */
  espece_ministerielle?: boolean;

  /** Default value: false */
  espece_cnpn?: boolean;
}

/** Represents the mutator for the table public.espece_protegee */
export interface EspeceProtegeeMutator {
  cd_ref?: EspeceProtegeeCdRef;

  classification?: string;

  noms_scientifiques?: string[];

  noms_vernaculaires?: string[];

  cd_type_statuts?: string[];

  espece_ministerielle?: boolean;

  espece_cnpn?: boolean;
}